port_open_cron: Make script self-referential so that it can be used with
certbot and other scripts. Now accepts "renew" to run the let's encrypt
renewal process, which in turn runs the pre_hook, renew_hook, and
post_hook hooks when appropriate. If not 'renew', then it simply runs
and evaluates the entire arguments list which I am using to run internal
functions.